pyproject.toml
[tool.poetry]
name = "cpc-project"
version = "0.1.0"
description = "CPC Practice Test Pipeline - Parse, Benchmark, and Fine-tune"
authors = ["natalie-a-1 <nataliehill1324@gmail.com>"]
readme = "README.md"
packages = [
    { include = "cpc_parser" },
    { include = "cpc_benchmark" },
    { include = "cpc_finetune" },
    { include = "cpc_tools" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
openai = "^1.0.0"
anthropic = "^0.55.0"
google-generativeai = "^0.8.5"
python-dotenv = "^1.0.0"
click = "^8.1.0"
tqdm = "^4.65.0"
jsonlines = "^4.0.0"
colorama = "^0.4.6"
pydantic = "^2.0.0"
tenacity = "^9.0.0"
pdfplumber = "^0.11.7"
faiss-cpu = "^1.7.0"
tiktoken = "^0.5.0"
loguru = "^0.7.0"
xlrd = "^2.0.0"
openpyxl = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.3.0"
mypy = "^1.16.1"
ipython = "^8.12.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=cpc_parser --cov=cpc_benchmark --cov=cpc_finetune --cov=cpc_tools --cov-report=term-missing"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
